Refactor user management: split user list into UserTable component, implement user fetching and filtering functionality
